Support HAL-FORMS encoding to GET requests
For HAL-FORMS templates with method set to GET (or HEAD); we can't send
any body, but need to send the request as query parameters instead.

We can only encode to the application/x-www-form-urlencoded content type
like this, because a query string is actually that encoding; but in the
URL instead of inside the request body.